requirements.txt
altgraph==0.16.1
Appium-Python-Client==0.28
browsermob-proxy==0.8.0
certifi==2019.9.11
chardet==3.0.4
ddt==1.4.4
demjson==2.2.4
Django==2.2.25
docopt==0.6.2
future==0.17.1
html-testRunner==1.2.1
idna==2.8
Jinja2==3.1.1
lxml==4.8.0
MarkupSafe==2.1.1
pefile==2019.4.18
phantomjs==1.4.1
pipreqs==0.4.11
PyInstaller==3.5
pytz==2021.3
pywin32==224
pywin32-ctypes==0.2.0
requests==2.22.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==3.14.0
sqlparse==0.4.2
urllib3==1.25.6
xlrd==1.2.0
yarg==0.1.9
